Refactor quest UI into smaller modules
